Create instruction
Hack Gmail with Social Engineering, you can apply this method to hack Facebook, Instagram accounts as well. But this tutorial is focused exclusively on how to hack Gmail accounts on Kali Linux with the Social Engineering Toolkit. But if you have just a little imagination, you can apply the same steps you will learn in this tutorial and duplicate the process to hack another account for another website as well. It is pretty much the same with just a few small variations.